Safari-611.1.10/LayoutTests/fast/dom/XMLSerializer-doctype2-expected.txt

This tests XMLSerializer.serializeToString() on a newly created DocumentType node does not throw since the node has an associated document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text = serializer.serializeToString(docType) did not throw exception.
PASS text is "<!DOCTYPE aDocTypeName PUBLIC \"aPublicID\" \"aSystemID\">"
